Added narration for player movement
This commit also makes the DungeonUniverse class more uniform
with how it returns a range of players (i.e., from a certain room,
or the whole list of players) by returning iterators over players.

Added for convenience with the DungeonDispatcher class is a method
that uses an iterator over players to place their PrintWriters into
an array, useful for the methods in the DungeonDispatcher class.

The Room class now has a getNumberOfPlayers() method.