Refactored all description and naming code
This commit beefs up the Describable class and transforms
it into an abstract class, containing default implementations
of methods that retrieve an ingame object's name and description.
All instances of previously used methods or redundant name/description
storages were replaced.

Also included in this commit are methods in the narrator class that
generate the names and descriptions for rooms, items, and players.
The methods were added here to make it more unified and translatable.
Uses of the old methods were refactored.

Also added was a flag in the Describable class that is used to prevent
any describable object's name from ever being mentioned with an article,
and a Gender class intended for use with languages that require it.