Refined how spaces are described
Removed override of inherited describe() method from the
Describable class in Space in favor of using the describe
method from the narrator class. Added a default toString method
to the narrator class, which uses no article.

Also changed the way the 'look' action behaves: when asked to
describe rooms, the server sends two notifications: one, the name
of the room (in all capitals) and the description on the second.