Initial work on universe file parsing and weather
Added to the universe class is a field that determines whether the
universe should generate and keep track of weather for the universe.

The major addition in this commit is the removal of the old default
universe and the addition of the universe file parsing code. The
server now checks for definitions of rooms and items in a YAML file
before starting the server. The format is described in part in the
"default.universe.yml" file, which is included here.